+++ packages/classic/src/hashes/index.ts
@@ -6,4 +6,4 @@ export { sha256 } from './sha256.js';
 export { keccak512, sha3 } from './sha3.js';
 export { sha384 } from './sha384.js';
 export { sha512 } from './sha512.js';
-export { registerClassicHashes } from './register.js';
+export { classicHashesPreset, registerClassicHashes } from './register.js';